+++ packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './validate';
 export * from './impl-store';
 export * from './version';
-export * from './manifest';
 export * from './storage';
 export * from './eip-1967';
 export * from './provider';
@@ -10,6 +9,15 @@ export * from './solc-api';
 export * from './deployment';
 export * from './link-refs';
 export * from './error';
+
+export {
+  ManifestData,
+  ImplDeployment,
+  ProxyDeployment,
+  Manifest,
+  migrateManifest,
+  DeploymentNotFound,
+} from './manifest';
 
 export { getStorageLayoutForAddress } from './manifest-storage-layout';