+++ packages/core/src/manifest.test.ts
@@ -1,5 +1,5 @@
 import test from 'ava';
-import { Manifest } from './manifest';
+import { Manifest, ManifestData, normalizeManifestData } from './manifest';
 
 test('manifest name for a known network', t => {
   const manifest = new Manifest(1);
@@ -10,4 +10,37 @@ test('manifest name for an unknown network', t => {
   const id = 55555;
   const manifest = new Manifest(id);
   t.is(manifest.file, `.openzeppelin/unknown-${id}.json`);
+});
+
+test('normalize manifest', t => {
+  const deployment = {
+    address: '0x1234',
+    txHash: '0x1234',
+    kind: 'uups' as const,
+    layout: { types: {}, storage: [] },
+    deployTransaction: {},
+  };
+  const input: ManifestData = {
+    manifestVersion: '3.0',
+    admin: deployment,
+    impls: { a: deployment },
+    proxies: [deployment],
+  };
+  const norm = normalizeManifestData(input);
+  t.like(norm.admin, {
+    ...deployment,
+    kind: undefined,
+    layout: undefined,
+    deployTransaction: undefined,
+  });
+  t.like(norm.impls.a, {
+    ...deployment,
+    kind: undefined,
+    deployTransaction: undefined,
+  });
+  t.like(norm.proxies[0], {
+    ...deployment,
+    layout: undefined,
+    deployTransaction: undefined,
+  });
 });